Record-only merge of r326857 and r326888
These are equal and opposite commits*; one to flip GELI support off by
default and another to flip it back on. They were made to circumvent some
GELI breakage in head that didn't apply to stable/11.

The former commit shows up when looking at MFC candidates for stand/, hence
the desire for this record.

* Equal and opposite except in that r326857 added a comment to zfsldr.S that
likely should not have been added at that time.


git-svn-id: https://svn.freebsd.org/base/stable/11@329151 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f