stand/lua: Correct some trivial errors in config
An empty module_path to start with isn't ideal, but if all modules are
contained within a kernel directory (which is what we just tested) then it
isn't strictly an error. Don't assume that module_path has a value already.

When we fail to load the kernel, printing the result (which is guaranteed to
be nil) is not intended; print the name of the kernel.


git-svn-id: https://svn.freebsd.org/base/head@329428 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f