Tweak the 'I' flagged value
'I' was omitting 'zero' values. This is not quite correct, and was put
in as a hack but not documented. Remove it. If we find what the hack
was really needed for, we'll either fix the need for it, or invent a
new flagged value type.

Submitted by: hps@
Sponsored by: Netflix


git-svn-id: https://svn.freebsd.org/base/head@329443 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f