stand/lua: Refactor logos into drawer.logodefs table
This refactor makes it straightforward to add new logos for drawing and
organizes logo definitions in a logical manner.

The graphic to be drawn for each logo may again be modified outside of
drawer, but it must be done on a case-by-case basis as a modification to the
loader_logo.


git-svn-id: https://svn.freebsd.org/base/head@329626 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f